data-prep/pyproject.toml
[project]
name = "data-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-groq>=0.3.2",
    "pydantic>=2.11.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "unidecode>=1.3.8",
]
